+++ docs/release_notes.rst
@@ -21,6 +21,8 @@ New Data Coverage
   <https://github.com/catalyst-cooperative/pudl-archiver/issues/296>`__.
 * Added new :ref:`core_eia860__scd_generators_wind` table from EIA860 Schedule 3.2
   which contains wind generator attributes. See :pr:`3522` and :pr:`3494`
+* Added new :ref:`core_eia860__scd_generators_solar` table from EIA860 Schedule 3.3
+  which contains solar generator attributes. See :pr:`3524` and :pr:`3482`
 
 Data Cleaning
 ^^^^^^^^^^^^^